refactor : 불필요 라인 삭제
