+++ cvs-batch/src/main/kotlin/com/yapp/cvs/job/crawl/WebdriverHandler.kt
@@ -10,7 +10,7 @@ interface WebdriverHandler {
     fun initializeWebdriver(): ChromeDriver {
         WebDriverManager.chromedriver().setup()
         val chromeOptions = ChromeOptions()
-        chromeOptions.setHeadless(true)
+//        chromeOptions.setHeadless(true)
         chromeOptions.addArguments("--remote-allow-origins=*")
         chromeOptions.addArguments("--no-sandbox", "--disable-dev-shm-usage")
         val driver = ChromeDriver(chromeOptions)
