+++ cvs-batch/src/main/kotlin/com/yapp/cvs/job/crawl/gs25/GS25CollectorConfig.kt
@@ -1,15 +1,15 @@
 package com.yapp.cvs.job.crawl.gs25
 
+import com.yapp.cvs.domain.collect.application.ProductDataProcessor
 import com.yapp.cvs.job.config.BatchConfig
 import com.yapp.cvs.job.crawl.gs25.handler.GS25WebdriverHandler
+import com.yapp.cvs.support.RunUniqueIdIncrementer
 import org.springframework.batch.core.Job
 import org.springframework.batch.core.Step
 import org.springframework.batch.core.configuration.annotation.JobBuilderFactory
 import org.springframework.batch.core.configuration.annotation.JobScope
 import org.springframework.batch.core.configuration.annotation.StepBuilderFactory
 import org.springframework.batch.core.configuration.annotation.StepScope
-import org.springframework.batch.core.launch.support.RunIdIncrementer
-import org.springframework.batch.core.repository.JobRepository
 import org.springframework.batch.core.step.tasklet.Tasklet
 import org.springframework.batch.support.transaction.ResourcelessTransactionManager
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty
@@ -23,8 +23,8 @@ import org.springframework.context.annotation.Configuration
 @Configuration
 class GS25CollectorConfig(
     private val jobBuilderFactory: JobBuilderFactory,
-    private val jobRepository: JobRepository,
     private val stepBuilderFactory: StepBuilderFactory,
+    private val productDataProcessor: ProductDataProcessor,
     private val webdriverHandler: GS25WebdriverHandler,
 ) {
     companion object {
@@ -35,9 +35,8 @@ class GS25CollectorConfig(
     @Bean
     fun gs25CollectorJob(): Job {
         return jobBuilderFactory[JOB_NAME]
-            .repository(jobRepository)
             .start(gs25CollectorStep())
-            .incrementer(RunIdIncrementer())
+            .incrementer(RunUniqueIdIncrementer())
             .build()
     }
 
@@ -46,6 +45,7 @@ class GS25CollectorConfig(
     fun gs25CollectorStep(): Step {
         return stepBuilderFactory[STEP_NAME]
             .tasklet(gs25CollectorTasklet())
+            .allowStartIfComplete(true)
             .transactionManager(ResourcelessTransactionManager())
             .build()
     }
@@ -54,5 +54,6 @@ class GS25CollectorConfig(
     @StepScope
     fun gs25CollectorTasklet(): Tasklet = GS25CollectorTasklet(
         webdriverHandler = webdriverHandler,
+        productDataProcessor = productDataProcessor,
     )
 }
