+++ cvs-batch/src/main/kotlin/com/yapp/cvs/job/crawl/gs25/GS25CollectorTasklet.kt
@@ -1,8 +1,9 @@
 package com.yapp.cvs.job.crawl.gs25
 
 import com.yapp.cvs.domain.collect.ProductRawDataVO
-import com.yapp.cvs.job.crawl.gs25.handler.GS25ProductCollectSupport
+import com.yapp.cvs.domain.collect.application.ProductDataProcessor
 import com.yapp.cvs.job.crawl.gs25.handler.GS25WebdriverHandler
+import com.yapp.cvs.support.GS25ProductCollectSupport
 import org.springframework.batch.core.StepContribution
 import org.springframework.batch.core.scope.context.ChunkContext
 import org.springframework.batch.core.step.tasklet.Tasklet
@@ -10,18 +11,26 @@ import org.springframework.batch.repeat.RepeatStatus
 
 class GS25CollectorTasklet(
     private val webdriverHandler: GS25WebdriverHandler,
+    private val productDataProcessor: ProductDataProcessor,
 ) : Tasklet {
     override fun execute(contribution: StepContribution, chunkContext: ChunkContext): RepeatStatus {
-        val driver = webdriverHandler.initializeWebdriver()
         val gs25ProductCollection = GS25ProductCollectSupport.values()
-        val productRawDataVOList = mutableListOf<ProductRawDataVO>()
-
-        gs25ProductCollection.forEach {
-            webdriverHandler.setCategoryTo(it, driver)
-            Thread.sleep(4000)
-            productRawDataVOList.addAll(webdriverHandler.collect(it, driver))
+        val driver = webdriverHandler.initializeWebdriver()
+        try {
+            gs25ProductCollection.forEach {
+                webdriverHandler.setCategoryTo(it, driver)
+                Thread.sleep(4000)
+                saveProductData(webdriverHandler.collect(it, driver))
+            }
+        } finally {
+            driver.quit()
         }
-        // todo : save all items
         return RepeatStatus.FINISHED
+    }
+
+    private fun saveProductData(productRawDataVOList: List<ProductRawDataVO>) {
+        productRawDataVOList.forEach {
+            productDataProcessor.saveProduct(it)
+        }
     }
 }
